Close admin dashboard full audit repair
